PROJECT3
I changed the name of the file because it is not what the teacher asked for exercise 8.I did not read instruction to the exercise 8 and I just continued working on the project itself using the exercise 7 base. But this is a version I want to use in the final project and will keep working on it after finishing exercise8.